% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energyr.R
\docType{package}
\name{energyr}
\alias{energyr}
\alias{energyr-package}
\title{An R package containing Federal Energy Regularity Commission (FERC) data}
\description{
energyr is an R package containing data from http://www.ferc.gov.
}
\section{energyr functions}{

electric
gas
hydropower
lng
oil
pipeline
storage
}

