% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gas.R
\docType{data}
\name{gas}
\alias{gas}
\title{Natural Gas Company Rates}
\format{Data frame with columns
\describe{
\item{Company}{Natural Gas Company}
\item{Revenue}{Annual Revenue in USD}
\item{Bill}{Annual Bill in USD}
\item{Year}{Reporting Year}
}}
\source{
http://www.ferc.gov/industries/gas/annual-charges.asp
}
\usage{
gas
}
\description{
Natural Gas Company Financial Information as reported by FERC.
}
\examples{
gas
}
\keyword{datasets}

