% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hydropower.R
\docType{data}
\name{hydropower}
\alias{hydropower}
\title{Hydropower Plant Information}
\format{Data frame with columns
\describe{
\item{Number}{FERC Hydropower Plant Number}
\item{Name}{Name of Hydropower Plant}
\item{Status}{Current Status of the Plant}
\item{Capacity}{Capacity of the Plant in KW}
\item{Company}{Name of the Hydropower Company}
\item{Waterway}{Name of the Waterway}
\item{lon}{Approximate Longitude of the Plant}
\item{lat}{Approximate Latitude of the Plant}
\item{Issued}{Issue Date of the License}
\item{Expiration}{Expiration Date of the License}
}}
\source{
http://www.ferc.gov/industries/hydropower.asp
}
\usage{
hydropower
}
\description{
Hydropower Plant Information as reported by FERC.
}
\examples{
hydropower
}
\keyword{datasets}

