% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lng.R
\docType{data}
\name{lng}
\alias{lng}
\title{LNG Plant Information}
\format{Data frame with columns
\describe{
\item{Location}{Location of the LNG Plant}
\item{Company}{Name of the LNG Company}
\item{Capacity}{Capacity of the Plant in Bcfd}
\item{Type}{Type of LNG Plant}
\item{Status}{Status of Plant}
\item{lat}{Approximate Latitude of the Plant}
\item{lon}{Approximate Longitude of the Plant}
}}
\source{
http://www.ferc.gov/industries/gas/indus-act/lng.asp
}
\usage{
lng
}
\description{
LNG Plant Information as reported by FERC.
}
\examples{
  dim(lng)
  head(lng)
}
\keyword{datasets}

