% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\docType{data}
\name{pipeline}
\alias{pipeline}
\title{Pipeline Project Information}
\format{
Data frame with columns
\describe{
\item{Name}{Name of the Pipeline Project}
\item{Operator}{Name of the Pipeline Operator}
\item{Type}{Type of Pipeline Project}
\item{Pipeline}{Type of Pipeline}
\item{Status}{Status of the Pipeline Project}
\item{Completed}{Date of Project Completion}
\item{Year}{Year of Project Completion}
\item{States}{State(s) of the Pipeline Project}
\item{Region}{FERC Region of the Pipeline Project}
\item{Cost}{Cost of the Pipeline Project in USD}
\item{Miles}{Miles of Pipeline}
\item{Capacity}{Capacity of the Pipeline in MMcf/d}
\item{Diameter}{Diameter(s) of the Pipeline in Inches}
\item{Pipeline}{Intrastate or Interstate Pipeline}
\item{Authority}{Regulatory Authority for the Project}
\item{Docket}{FERC Docket Number}
}
}
\source{
http://www.ferc.gov/industries/gas/indus-act/pipelines.asp
}
\usage{
pipeline
}
\description{
Natural Gas Pipeline Project Information as reported by FERC.
}
\examples{
  dim(pipeline)
  head(pipeline)
}
\keyword{datasets}
