% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/storage.R
\docType{data}
\name{storage}
\alias{storage}
\title{Natural Gas Storage Field Information}
\format{Data frame with columns
\describe{
\item{lon}{Approximate Longitude of the Field}
\item{lat}{Approximate Latitude of the Field}
\item{Company}{Name of the Storage Company}
\item{Field}{Name of the Field}
\item{Location}{Location of the Field}
\item{Type}{Type of the Field}
\item{Total}{Total Capacity in Mcf}
\item{Working}{Working Capacity in Mcf}
}}
\source{
http://www.ferc.gov/industries/gas/indus-act/lng.asp
}
\usage{
storage
}
\description{
Natural Gas Storage Field Information as reported by FERC.
}
\examples{
storage
}
\keyword{datasets}

