% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/electric.R
\docType{data}
\name{electric}
\alias{electric}
\title{Electric Company Rates}
\format{Data frame with columns
\describe{
\item{Company}{Electric Company}
\item{Revenue}{Annual Revenue in USD}
\item{Bill}{Annual Bill in USD}
\item{Year}{Reporting Year}
}}
\source{
http://www.ferc.gov/industries/electric/annual-charges.asp
}
\usage{
electric
}
\description{
Electric Company Financial Information as reported by FERC.
}
\examples{
  dim(electric)
  head(electric)
}
\keyword{datasets}

