% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydropower.R
\docType{data}
\name{hydropower}
\alias{hydropower}
\title{Hydropower Plant Information}
\format{
Data frame with columns
\describe{
\item{Number}{FERC Hydropower Plant Number}
\item{Name}{Name of Hydropower Plant}
\item{Expiration}{Expiration Date of the License}
\item{Issued}{Issue Date of the License}
\item{Status}{Current Status of the Plant}
\item{Capacity}{Capacity of the Plant in KW}
\item{Company}{Name of the Hydropower Company}
\item{Waterway}{Name of the Waterway}
\item{lat}{Approximate Latitude of the Plant}
\item{lon}{Approximate Longitude of the Plant}
}
}
\source{
http://www.ferc.gov/industries/hydropower.asp
}
\usage{
hydropower
}
\description{
Hydropower Plant Information as reported by FERC.
}
\examples{
  dim(hydropower)
  head(hydropower)
}
\keyword{datasets}
